Windows port: transformer: fix MSVC narrowing and GNU-builtin errors
- sdpa/device/exp_ring_joint_sdpa_program_factory.cpp: static_cast size_t
  ring_size/logical_n/q_chunk_start values to uint32_t in braced kernel
  compile-time-arg vectors and struct initializers (C2397/C2398)
- sdpa/device/ring_joint_sdpa_device_operation.cpp: static_cast size_t
  num_devices to uint32_t when building RingAttentionAllGatherAsyncParams
  (C2398)
- sdpa/device/ring_joint_sdpa_program_factory.cpp: static_cast size_t
  grid_size.x arithmetic to uint32_t for num_receivers (C2397)
- sdpa_decode/device/sdpa_decode_device_operation.hpp,
  sdpa_decode_program_factory.cpp: replace __builtin_clz/__builtin_ctz
  with the portable tt::compiler::count_leading_zeros32/
  count_trailing_zeros32 helpers from hostdevcommon/tt_compiler.h
  (C3861, plus the C2737 this masked)

+++ ttnn/cpp/ttnn/operations/transformer/sdpa/device/ring_joint_sdpa_device_operation.cpp
@@ -1239,7 +1239,7 @@ RingJointSDPAResult ring_joint_scaled_dot_product_attention(
         {},
         gather_dim,
         num_links,
-        num_devices,
+        static_cast<uint32_t>(num_devices),
         persistent_output_buffer_k.memory_config(),
         topology,
         multi_device_global_semaphore,
